requirements.txt
aiogram==3.4.1
aiosqlite==0.19.0
python-dotenv==1.0.0
pytz==2024.1
apscheduler==3.10.4

